refactor(agent): split the summary-plan helpers and trim lint overruns
